Add image and content on lasagna recipe page
Added illustrative image of lasagna in 'lasagna.html'.
Added a "Description" title to 'lasagna.html'.
Included a paragraph of recipe description in 'lasagna.html'.
Included an unordered list of required ingredients in 'lasagna.html'.
Added a "Steps" title to 'lasagna.html'.
Included an ordered list of recipe steps in 'lasagna.html'.

This commit populates the content on the lasagna recipe page and adds a lasagna image.